build(.* files, script files): Adding all the standard .* files, dependency files and script tooling
The .babelrc file is for transpiling javascript, if necessary. .eslintignore and .eslintrc tell
ESlint to not lint some files and the rules to use when linting, respectively. .gitignore tells
github what files not to commit. .opt-in forces github hooks to execute before a commit is made to
keep commit messages standard and semantic-release working correctly. package-lock.json and
package.json is NPM's dependency management and package-scripts.js is a file used by a build tool
nps to keep scripts that run testing, building neat.