test(cli): Added integration tests for cli and culebra.js
Added two separate integration tests for culebra and index.js. The reason is that one tests the cli
functionality and the other the request functionality. I broke it up in to two similar test because
the manner in which we test the cli does not allow for the coverage module to take effect. By doing
them separetely. It takes appropriate coverage metrics.